split datapoints tests
